spread event listeners
